Ensure serialized size_t can be passed between 32bit and 64bit processes
https://crrev.com/c/4205862 broken some WebView environment which
uses 32bit renderer process and 64bit browser process because size_t
was serialized as the native type uint32_t or uint64_t.

Now partly revert the CL about size_t to always serialize size_t as
uint64_t.

The current PaintOpReader and PaintOpWriter are still fragile because
the caller can use Write/Read() instead of WriteSize()/ReadSize() for
size_t and cause problems. I will follow up to refactor this to prevent
the problem.

Bug: 1414445, 1414686
Cq-Include-Trybots: luci.chromium.try:android-webview-pie-arm64-dbg,android-webview-oreo-arm64-dbg
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4237958
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1104172}

+++ cc/paint/paint_op_writer.cc
@@ -176,7 +176,7 @@ void PaintOpWriter::WriteFlattenable(const SkFlattenable* val) {
     return;
   }
 
-  size_t* size_memory = WriteSize(0u);
+  uint64_t* size_memory = WriteSize(0u);
   if (!valid_)
     return;
 
@@ -190,9 +190,11 @@ void PaintOpWriter::WriteFlattenable(const SkFlattenable* val) {
   DidWrite(bytes_written);
 }
 
-size_t* PaintOpWriter::WriteSize(size_t size) {
-  size_t* memory = reinterpret_cast<size_t*>(memory_.get());
-  WriteSimple(size);
+uint64_t* PaintOpWriter::WriteSize(size_t size) {
+  // size_t is always serialized as uint64_t to make the serialized result
+  // portable between 32bit and 64bit processes.
+  uint64_t* memory = reinterpret_cast<uint64_t*>(memory_.get());
+  WriteSimple(static_cast<uint64_t>(size));
   return memory;
 }
 
@@ -257,7 +259,7 @@ void PaintOpWriter::Write(const SkPath& path, UsePaintCache use_paint_cache) {
   } else {
     Write(static_cast<uint32_t>(PaintCacheEntryState::kInlinedDoNotCache));
   }
-  size_t* bytes_to_skip = WriteSize(0u);
+  uint64_t* bytes_to_skip = WriteSize(0u);
   if (!valid_)
     return;
 
@@ -350,7 +352,7 @@ void PaintOpWriter::Write(scoped_refptr<SkottieWrapper> skottie) {
   uint32_t id = skottie->id();
   Write(id);
 
-  size_t* bytes_to_skip = WriteSize(0u);
+  uint64_t* bytes_to_skip = WriteSize(0u);
   if (!valid_)
     return;
 
@@ -454,7 +456,7 @@ void PaintOpWriter::Write(const sk_sp<GrSlug>& slug) {
     return;
 
   AssertFieldAlignment();
-  size_t* size_memory = WriteSize(0u);
+  uint64_t* size_memory = WriteSize(0u);
   if (!valid_)
     return;
 
@@ -972,7 +974,7 @@ void PaintOpWriter::Write(const PaintRecord& record,
   // We need to record how many bytes we will serialize, but we don't know this
   // information until we do the serialization. So, write 0 as the size first,
   // and amend it after writing.
-  size_t* size_memory = WriteSize(0u);
+  uint64_t* size_memory = WriteSize(0u);
   if (!valid_)
     return;
